requirements.txt
httpx>=0.26.0
typer>=0.9.0
rich>=13.7.0
python-dotenv>=1.0.0